requirements.txt
# Sistema de Rehabilitación Motriz - Visión por Computadora
# Dependencias para prototipo y scripts de análisis

opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
fpdf2>=2.7.0
pydantic>=2.0.0
dearpygui>=1.11.0
